refactor: split components for the draw character page
